#93, #95, #96 and #97 compilation problems on bluepill and mbed, changes to Java UI to come in line with Store version. Non negative large numbers

+++ CoreXmlPlugins/core-display/lcdSrc/tcMenuLiquidCrystal.cpp
@@ -82,7 +82,7 @@ void LiquidCrystalRenderer::renderTitle(bool forceDraw) {
         strcpy_P(buffer, applicationInfo.name);
         serdebugF2("print app name", buffer);
         uint8_t bufSz = bufferSize;
-        uint8_t last = min(bufSz, strlen(buffer));
+        uint8_t last = min(bufSz, (uint8_t)strlen(buffer));
         for(uint8_t i = last; i < bufSz; i++) {
             buffer[i] = ' ';
         }
